pyproject.toml
[project]
name = "func-thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.24.0",
    "azure-identity>=1.25.1",
    "azure-keyvault-secrets>=4.10.0",
    "azure-storage-blob>=12.27.0",
    "httpx>=0.28.1",
    "pytest>=9.0.0",
]
